fix: header uses long int change variable to long int

+++ src/utils.c
@@ -3,11 +3,11 @@
 #include <time.h>
 
 
-void fmt_centavos(int centavos, char *buf, int bufsz)
+void fmt_centavos(int64_t centavos, char *buf, int bufsz)
 {
     int negative = (centavos < 0);
     if (negative) centavos = -centavos;
-    snprintf(buf, (size_t)bufsz, "%sPHP %d.%02d",
+    snprintf(buf, (size_t)bufsz, "%sPHP %ld.%02d",
              negative ? "-" : "",
              centavos / 100,
              centavos % 100);
